requirements.txt
asgiref==3.8.1
Django==5.1.2
djangorestframework==3.15.2
psycopg2==2.9.10
psycopg2-binary==2.9.10
python-decouple==3.8
sqlparse==0.4.3
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
